+++ android/src/main/java/vn/map4d/map/map4d_map/FMFPOIBuilder.java
@@ -53,7 +53,7 @@ class FMFPOIBuilder implements FMFPOIOptionsSink {
 
   @Override
   public void setTitleColor(int titleColor) {
-    poiOptions.titleColor(titleColor);
+    poiOptions.color(titleColor);
   }
 
   @Override