+++ example/lib/map_main.dart
@@ -1,10 +1,6 @@
-import 'dart:math';
-
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
 
-import 'package:flutter/services.dart';
 import 'package:map4d_map/map4d_map.dart';
 import 'page.dart';
 
@@ -26,7 +22,6 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   bool _isMapCreated = false;
-  bool _is3DMode = false;
   late MFMapViewController _controller;
 
   @override
@@ -47,7 +42,6 @@ class _MyAppState extends State<MyApp> {
       myLocationEnabled: true,
       myLocationButtonEnabled: true,
       onTap: onTap,
-      onModeChange: onModeChange,
       onPOITap: onBaseMapPOITap,
       onBuildingTap: onBaseMapBuildingTap,
       onPlaceTap: onBaseMapPlaceTap,
@@ -146,8 +140,6 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> doAction() async {
-    // _is3DMode = !_is3DMode;
-    // _controller.enable3DMode(_is3DMode);
     // const p = MFScreenCoordinate(x: 100, y: 200);
     // _controller.getLatLng(p).then((latlng) {
     //   print('Convert point ${p.toJson()} to latlng: ${latlng.toJson()}');
@@ -168,14 +160,6 @@ class _MyAppState extends State<MyApp> {
 
   void onTap(MFLatLng coordinate) {
     print('Did tap ' + coordinate.toString());
-  }
-
-  void onModeChange(bool is3Dmode) {
-    var mode = '2D';
-    if (is3Dmode) {
-      mode = '3D';
-    }
-    print('Mode of map is: ' + mode);
   }
 
   void onBaseMapPOITap(String placeId, String name, MFLatLng location) {