+++ lib/src/mapview.dart
@@ -32,6 +32,7 @@ class MFMapView extends StatefulWidget {
     this.onMapCreated,
     this.gestureRecognizers = const <Factory<OneSequenceGestureRecognizer>>{},
     this.mapId,
+    this.style,
     this.mapType = MFMapType.roadmap,
     this.minMaxZoomPreference = MFMinMaxZoom.unbounded,
     this.rotateGesturesEnabled = true,
@@ -46,8 +47,6 @@ class MFMapView extends StatefulWidget {
     this.onCameraMove,
     this.onCameraIdle,
     this.onTap,
-    @Deprecated('This method is no longer support, which is subject to removal in a future versions.')
-        this.onModeChange,
     this.onPOITap,
     this.onBuildingTap,
     this.onPlaceTap,
@@ -114,6 +113,9 @@ class MFMapView extends StatefulWidget {
   /// An opaque identifier for a custom map configuration.
   final String? mapId;
 
+  /// The locally configured style for the map.
+  final String? style;
+
   /// Type of map tiles to be rendered.
   final MFMapType mapType;
 
@@ -155,11 +157,6 @@ class MFMapView extends StatefulWidget {
 
   /// Called when did tap at coordinate
   final MFLatLngCallback? onTap;
-
-  /// Called when map mode change from 2D -> 3D & 3D -> 2D
-  @Deprecated(
-      'This method is no longer support, which is subject to removal in a future versions.')
-  final MFModeChangedCallback? onModeChange;
 
   /// Called after a base map POI has been tapped
   final MFMapPOICallback? onPOITap;
@@ -473,12 +470,14 @@ class _MFMapViewOptions {
         buildingsEnabled = map.buildingsEnabled,
         poisEnabled = map.poisEnabled,
         mapId = map.mapId,
+        style = map.style,
         mapType = map.mapType,
         minMaxZoomPreference = map.minMaxZoomPreference,
         trackCameraPosition = map.onCameraMove != null;
 
   // final CameraTargetBounds cameraTargetBounds;
   final String? mapId;
+  final String? style;
   final MFMapType mapType;
   final MFMinMaxZoom minMaxZoomPreference;
   final bool rotateGesturesEnabled;
@@ -510,6 +509,10 @@ class _MFMapViewOptions {
 
     if (mapId != null) {
       map["mapId"] = mapId;
+    }
+
+    if (style != null) {
+      map["style"] = style;
     }
 
     return map;