+++ lib/src/ui.dart
@@ -1,16 +1,13 @@
 /// Type of map tiles to display
 enum MFMapType {
-  /// Roadmap maps, default
+  /// Roadmap tiles, default
   roadmap,
 
-  /// Raster maps
-  raster,
-
-  /// Satellite maps
+  /// Satellite imaging tiles
   satellite,
 
-  /// 3D maps
-  map3D,
+  /// Hybrid tiles
+  hybrid,
 }
 
 class MFMinMaxZoom {